requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==24.2
pillow==11.1.0
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
uritemplate==4.1.1
